Moved BattleRepository to subpackage
 Changes to be committed:
	renamed:    src/main/java/com/ardaslegends/repository/BattleRepository.java -> src/main/java/com/ardaslegends/repository/war/battle/BattleRepository.java
	modified:   src/main/java/com/ardaslegends/service/war/BattleService.java
	modified:   src/test/java/com/ardaslegends/service/BattleServiceTest.java
