+++ src/main/java/com/ardaslegends/service/war/BattleService.java
@@ -3,9 +3,9 @@ package com.ardaslegends.service.war;
 import com.ardaslegends.domain.*;
 import com.ardaslegends.domain.war.battle.Battle;
 import com.ardaslegends.domain.war.battle.BattleLocation;
-import com.ardaslegends.repository.*;
 import com.ardaslegends.repository.war.WarRepository;
 import com.ardaslegends.repository.war.QueryWarStatus;
+import com.ardaslegends.repository.war.battle.BattleRepository;
 import com.ardaslegends.service.*;
 import com.ardaslegends.service.dto.war.ConcludeBattleDto;
 import com.ardaslegends.service.dto.war.battle.CreateBattleDto;
